refactor: 동일 부류 dto 패키징
