+++ src/main/kotlin/lotto/view/dto/lotto/LottoDto.kt
@@ -1,3 +1,3 @@
-package lotto.view.dto
+package lotto.view.dto.lotto
 
 data class LottoDto(val numbers: List<Int>)
