+++ src/main/kotlin/lotto/view/dto/lotto/LottosDto.kt
@@ -1,3 +1,3 @@
-package lotto.view.dto
+package lotto.view.dto.lotto
 
 data class LottosDto(val lottos: List<LottoDto>)
